mv scattergl-only tests in gl2d_plot_interact_test.js
... in scattergl_test.js and scattergl_select_test.js
